Writing an And/or Tree model.
The And/Or Tree model will be integrated with the language comprehension program. I am still not sure how the tree would be in the end, but it might be closer to a neural net but functions like an And/Or Tree.

I won't make any update until next Tuesday after coming back from camping trip.